requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
passlib
bcrypt
python-jose
python-multipart
